crawler/requirements.txt
# 화장품 상담 시스템 크롤러 의존성

# 웹 크롤링
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2
webdriver-manager==4.0.1

# 데이터베이스
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# 스케줄링
schedule==1.2.1
python-crontab==3.0.0

# 유틸리티
python-dotenv==1.0.0
pytz==2023.3
python-dateutil==2.8.2

# 로깅 및 모니터링
colorlog==6.8.0

# HTTP 클라이언트
httpx==0.25.2

# HTML 파싱
html5lib==1.1

# 데이터 처리
pandas==2.1.4

# UUID 생성
uuid==1.30


